+++ lib/registry.ts
@@ -24,7 +24,7 @@ import {
 
 import { maybeRetrievePositionFromSourceMap, Position } from "./source-map";
 
-interface IStepDefinition<T extends unknown[]> {
+export interface IStepDefinition<T extends unknown[]> {
   expression: Expression;
   implementation: IStepDefinitionBody<T>;
   position?: Position;
@@ -62,7 +62,7 @@ function parseHookArguments(
 }
 
 export class Registry {
-  private parameterTypeRegistry: ParameterTypeRegistry;
+  public parameterTypeRegistry: ParameterTypeRegistry;
 
   private preliminaryStepDefinitions: {
     description: string | RegExp;
@@ -112,10 +112,6 @@ export class Registry {
   }
 
   public defineStep(description: string | RegExp, implementation: () => void) {
-    let position: Position | undefined;
-
-    position = maybeRetrievePositionFromSourceMap(this.experimentalSourceMap);
-
     if (typeof description !== "string" && !(description instanceof RegExp)) {
       throw new Error("Unexpected argument for step definition");
     }
@@ -123,7 +119,7 @@ export class Registry {
     this.preliminaryStepDefinitions.push({
       description,
       implementation,
-      position,
+      position: maybeRetrievePositionFromSourceMap(this.experimentalSourceMap),
     });
   }
 
@@ -159,10 +155,14 @@ export class Registry {
     );
   }
 
-  private resolveStepDefintion(text: string) {
-    const matchingStepDefinitions = this.stepDefinitions.filter(
-      (stepDefinition) => stepDefinition.expression.match(text)
+  public getMatchingStepDefinitions(text: string) {
+    return this.stepDefinitions.filter((stepDefinition) =>
+      stepDefinition.expression.match(text)
     );
+  }
+
+  public resolveStepDefintion(text: string) {
+    const matchingStepDefinitions = this.getMatchingStepDefinitions(text);
 
     if (matchingStepDefinitions.length === 0) {
       throw new MissingDefinitionError(